requirements.txt
Flask
Jinja2
MarkupSafe
Werkzeug
argparse
itsdangerous
wsgiref
requests
python-firebase